requirements.txt
fastapi
uvicorn[standard]
sqlmodel
instructor
openai
python-dotenv